requirements.txt
fastapi>=0.110.0,<1.0
uvicorn[standard]>=0.27.0,<1.0
jinja2>=3.1.2,<4.0
starlette>=0.36.0,<1.0
pydantic>=2.6.0,<3.0
numpy>=1.24.0,<2.0
# CPU-only TensorFlow is smaller for typical deployments
tensorflow-cpu>=2.12.0,<3.0
